requirements.txt
requests
pytest
allure-pytest
Faker
